mission-done

+++ server/app/Http/Controllers/Api/MemberController.php
@@ -13,6 +13,9 @@ use Illuminate\Support\Facades\DB;
 class MemberController extends Controller
 {
     public function show($projectId, $id) {
+        $project = Project::find($projectId);
+        $this->authorize('view', $project);
+
         $member = User::find($id);
         $role = Role::where('user_id', $id)->where('project_id', $projectId)->get();
 
@@ -22,6 +25,8 @@ class MemberController extends Controller
     public function edit($id)
     {
         $project = Project::find($id);
+        $this->authorize('view', $project);
+
         $users = $project->users()->get();
         $roles = $project->roles;
 
@@ -29,10 +34,13 @@ class MemberController extends Controller
     }
 
     public function create($projectId) {
+        $project = Project::find($projectId);
+        $this->authorize('view', $project);
+
         $users = User::whereDoesntHave('projects', function ($query) use ($projectId){
             $query->where('project_id', $projectId);
         })->get();
-        $project = Project::find($projectId);
+
         $roles = $project->roles;
 
         return response()->json(['users' => $users, 'roles' => $roles]);
@@ -40,6 +48,8 @@ class MemberController extends Controller
 
     public function store($id) {
         $project = Project::findOrFail($id);
+        $this->authorize('view', $project);
+
         $role = Role::find(request()->input('role_id'));
 
         Role::create([
@@ -52,6 +62,9 @@ class MemberController extends Controller
     }
 
     public function update($memberId, $roleId, $projectId) {
+        $project = Project::find($projectId);
+        $this->authorize('update', $project);
+
         DB::table('roles')
             ->where('project_id', $projectId)
             ->where('user_id', $memberId)
@@ -63,6 +76,9 @@ class MemberController extends Controller
     }
 
     public function destroy($projectId, $memberId) {
+        $project = Project::find($projectId);
+        $this->authorize('delete', $project);
+
         $user = User::find($memberId);
         Message::where('project_id', $projectId)->where('sender_id', $memberId)->delete();
         Role::where('project_id', $projectId)->where('user_id', $memberId)->update(['user_id' => 0]);
